+++ module-check/src/test/clojure/clojure/core/typed/test/swap_bang.clj
@@ -7,6 +7,11 @@
 (fn []
   (swap! foo assoc :a 3))
 
+(t/ann b (t/Atom1 (t/Vec String)))
+(def b (atom []))
+
+(swap! b (fn [in] in))
+
 ;
 ;(swap! foo (fn [a] a))
 ;(swap! foo (fn [a] a))